+++ src/eval/bc/stencils/loop_int_dec_inc.c
@@ -36,7 +36,14 @@ void stencil_op_loop_int_dec_inc(mino_val_t **regs,
                                   mino_val_t **consts,
                                   mino_state_t *S)
 {
+    unsigned long ticks = 256;
     for (;;) {
+        if (__builtin_expect(--ticks == 0, 0)) {
+            if (!mino_bc_safepoint(S)) {
+                MINO_STENCIL_CHAIN_RETURN(NULL, consts, S);
+            }
+            ticks = 256;
+        }
         mino_val_t *vt = regs[IMM_A];
         mino_val_t *vi = regs[IMM_B];
         uintptr_t ut = (uintptr_t)vt;